pyproject.toml
[tool.poetry]
name = "prometheus-grafana-loki-api-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
requests = "^2.32.3"
pytz = "^2024.2"
httpx = "^0.27.2"
python-logging-loki = "^0.3.1"
uvicorn = "^0.31.1"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
